Move the rbac functions into their component source files
These only apply to a single component and are only called from
that component function in the controller. It's strange that they're
in a separate source file.
